backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
psycopg2-binary>=2.9.11
python-decouple==3.8
django-cors-headers==4.3.1
bcrypt==4.1.2
gunicorn==21.2.0
cloudinary==1.41.0
Pillow==10.4.0

